% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bicuar}
\alias{bicuar}
\title{Dataframe of example stem data}
\format{
An object of class \code{data.frame} with 1313 rows and 6 columns.
}
\usage{
data(bicuar)
}
\description{
Dataframe of stem data from Bicuar National Park, in the SEOSAW style, but
only with columns relevant to the functions in the \code{{ProdVital}} package.
}
\author{
John L. Godlee
}
\keyword{datasets}
