% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bicuar_clean}
\alias{bicuar_clean}
\title{Dataframe of cleaned example stem data}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2287 rows and 10 columns.
}
\usage{
data(bicuar_clean)
}
\description{
Dataframe of stem data from Bicuar National Park, in the SEOSAW style, but
only with columns relevant to the functions in the \code{{ProdVital}} package.
Used in examples. The same as \code{data(bicuar)} after cleaning as laid out in
the vignette.
}
\author{
John L. Godlee
}
\keyword{datasets}
