% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthAll.R
\name{growthAll}
\alias{growthAll}
\title{Calculate growth increment for consecutive or pairwise intervals}
\usage{
growthAll(x, w, group, census, type = "consecutive")
}
\arguments{
\item{x}{dataframe of SEOSAW stem data}

\item{w}{column name of value from which to calculate growth in \code{x}}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}

\item{type}{string defining whether to return only consecutive or all pairwise census intervals.}
}
\value{
Dataframe of individual growth increment values across all
consecutive or pairwise census intervals.

Columns:
\itemize{
\item \code{t0} - date of the start of the census interval
\item \code{tT} - date of the end of the census interval
\item \code{w0} - The value of \code{w} at the start of the census interval
\item \code{wT} - The value of \code{w} at the end of the census interval
\item \code{g} - Growth increment (\code{wT - w0})
}
}
\description{
Calculate growth increment for consecutive or pairwise intervals
}
\details{
This function calculates growth increments between census
intervals. Column \code{t0} is the initial census in the census
interval, \code{tT} is the final census, \code{int} is the length of the
census, and \code{g} is the growth increment.

\code{type} must be one of "consecutive" or "pairwise".

Only individuals with >1 census are returned.
}
\examples{
data(bicuar_clean)

growthAll(bicuar_clean, w = "diam",
  group = c("plot_id", "stem_id"), census = "census_date",
  type = "consecutive")

growthAll(bicuar_clean, w = "diam",
  group = c("plot_id", "stem_id"), census = "census_date",
  type = "pairwise")

}
