% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthMod.R
\name{growthMod}
\alias{growthMod}
\title{Estimate average growth rate of individuals}
\usage{
growthMod(x, w, group, census, full = FALSE)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data}

\item{w}{column name of value from which to calculate growth in \code{x}}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}

\item{full}{logical, if \code{TRUE} return a dataframe, otherwise a named vector}
}
\value{
Named vector or dataframe (if \code{full = TRUE}) of individual observed average growth rates across all census
intervals.
}
\description{
Estimate average growth rate of individuals
}
\details{
This function fits a linear model and returns the slope of that
model, which is identical to the average rate of growth, in units
of \code{w} per \code{census}.

Only individuals with >1 census are returned.
}
\examples{
data(bicuar_clean)

growthMod(bicuar_clean, w = "diam",
  group = c("plot_id", "stem_id"), census = "census_date")

growthMod(bicuar_clean, w = "diam",
  group = c("plot_id", "stem_id"), census = "census_date", 
  full = TRUE)

}
