% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indFill.R
\name{indFill}
\alias{indFill}
\title{Create missing individual records}
\usage{
indFill(x, d, group, census)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data}

\item{d}{vector of possible census dates for individuals in \code{x}}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}
}
\value{
dataframe of imputed individual records
}
\description{
Create missing individual records
}
\details{
\code{group} can be a vector of multiple character strings which, when combined, uniquely define each individual.
}
\examples{
x <- data.frame(
  stem_id = "1",
  census_date = c(2000, 2002))
d = c(2000, 2001, 2002)
indFill(x, d, group = "stem_id", census = "census_date")

}
