% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indGrowth.R
\name{indGrowth}
\alias{indGrowth}
\title{Calculate growth increments}
\usage{
indGrowth(x, t0, tT, w, group, census)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data}

\item{t0}{initial census of interval}

\item{tT}{final census of interval}

\item{w}{column name of value from which to calculate growth in \code{x}}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}
}
\value{
Named vector of individual growth increment. Names are the concatenation
(sep = \code{.}) of all grouping columns provided in \code{group}.
}
\description{
This function acts on a datafame (\code{x}) of individuals (\code{group}) measured at two time points (\code{t0}, \code{tT}, \code{census}) to calculate the growth increment of each individual
given a measure of individual size (\code{w}).
}
\details{
\code{group} can be a vector of multiple character strings which, when combined, uniquely define each individual.

\code{t0} and \code{tT} by default are years.
}
\examples{
data(bicuar)

indGrowth(bicuar, t0 = "2019", tT = "2021", w = "agb", 
  group = "stem_id", census = "census_date")

}
