% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortFill.R
\name{mortFill}
\alias{mortFill}
\title{Impute missing and inconsistent mortality status}
\usage{
mortFill(x, group, census, status)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}

\item{status}{column name of census dates in \code{x}}
}
\value{
numeric vector of length \code{nrow(x)} with estimated individual
mortality status (1 = alive, 0 = dead)
}
\description{
Impute missing and inconsistent mortality status
}
\details{
Missing mortality status values are interpolated where possible. If the
missing value(s) is both preceded and followed by the same mortality status,
this value is used to fill the missing value(s). If the missing value(s) is
the first or final census(es) and either directly preceded or followed by two or more
of the same mortality status, this value is used to fill the missing
value(s). After these steps, all values occurring before the final alive
value are changed to alive. \code{group} can be a vector of multiple character strings which, when combined, uniquely define each individual.
}
\examples{
status_list <- list(
  c(NA_real_, NA_real_, 0, 0),
  c(NA_real_, NA_real_, 1, 0),
  c(NA_real_, NA_real_, 1, 1),
  c(NA_real_, NA_real_, 1),
  c(1, NA_real_, 1, 0),
  c(1, NA_real_, NA_real_, 0),
  c(1, NA_real_, NA_real_, 0),
  c(1, 0, 1, 0),
  c(1, NA_real_, 1, NA_real_))

lapply(status_list, function(x) { 
  d <- data.frame(
    stem_id = "a", 
    census_date = seq(2000, length.out = length(x)),
    status = x)

  mortFill(d, group = "stem_id",
    census = "census_date", status = "status")
})

}
