% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsID.R
\name{obsID}
\alias{obsID}
\title{Find individuals which survived, died, or recruited}
\usage{
obsID(x, t0, tT, type = c("rec", "sur", "mor"), group, census)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data}

\item{t0}{initial census of interval}

\item{tT}{final census of interval}

\item{type}{vector defining which individuals to return, either: \code{rec} for recruits, \code{sur} for survivors, or \code{mor} for deaths}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}
}
\value{
Dataframe or list of dataframes of IDs for individuals which recruited,
survived, or died, between censuses.
}
\description{
This function acts on a datafame (\code{x}) of individuals (\code{group}) measured at two time points (\code{t0}, \code{tT}, \code{census}) to return the identities of individuals which survived
(measured at both time points), recruited (measured only at the final time
point), or died (measured only at the initial time point).
}
\details{
\code{group} can be a vector of multiple character strings which, when combined, uniquely define each individual.
}
\examples{
data(bicuar_clean)

obsID(bicuar_clean, "2019", "2021", 
  group = c("plot_id", "stem_id"), census = "census_date")

obsID(bicuar_clean, "2019", "2021", type = "mor", 
  group = c("plot_id", "stem_id"), census = "census_date")

}
