% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsMorLoss.R
\name{obsMorLoss}
\alias{obsMorLoss}
\title{Calculate loss from individuals which died}
\usage{
obsMorLoss(x, t0, tT, w, group, census, full = FALSE)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data}

\item{t0}{initial census of interval}

\item{tT}{final census of interval}

\item{w}{column name of value from which to calculate growth in \code{x}}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}

\item{full}{logical, if \code{TRUE} return a dataframe, otherwise a named vector}
}
\value{
Named vector or dataframe (if \code{full = TRUE}) of individual observed loss from individuals which died.
}
\description{
This function acts on a datafame (\code{x}) of individuals (\code{group}) measured at two time points (\code{t0}, \code{tT}, \code{census}) to calculate the loss from individuals which died, given a measure of individual size (\code{w}).
}
\details{
\code{group} can be a vector of multiple character strings which, when combined, uniquely define each individual.
}
\examples{
data(bicuar_clean)

obsMorLoss(bicuar_clean, "2019", "2021", w = "agb", 
  group = c("plot_id", "stem_id"), census = "census_date")
 
}
