% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsRecGrowth.R
\name{obsRecGrowth}
\alias{obsRecGrowth}
\title{Estimate growth of individuals which recruited between two censues}
\usage{
obsRecGrowth(
  x,
  t0,
  tT,
  w,
  group,
  census,
  rec_method = "zero",
  diam = NULL,
  min_size_class = NULL,
  min_diam_thresh = NULL,
  growth_percentile = NULL,
  w_min_diam = NULL
)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data}

\item{t0}{initial census of interval}

\item{tT}{final census of interval}

\item{w}{column name of value from which to calculate growth in \code{x}}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}

\item{rec_method}{character string describing method used to estimate growth of recruits. Can be \code{zero}, \code{thresh}, or \code{extrap}.}

\item{diam}{optional, column name of stem diameters in \code{x}}

\item{min_size_class}{optional, vector of length two containing range of minimum diameter size class used to estimate median growth rate}

\item{min_diam_thresh}{optional, minimum diameter threshold measured}

\item{growth_percentile}{optional, percentile of growth rate used to estimate growth rate of recruits for backward extrapolation}

\item{w_min_diam}{optional, column name or a vector of length 1 defining the estimated value of \code{w} in \code{x} at the minimum diameter threshold of the plot}
}
\value{
Numeric value of the sum of all observed growth from individuals which recruited.
}
\description{
This function acts on a datafame (\code{x}) of individuals (\code{group}) measured at two time points (\code{t0}, \code{tT}, \code{census}) to calculate the growth from individuals which recruited, given a measure of individual size (\code{w})
}
\details{
\code{group} can be a vector of multiple character strings which, when combined, uniquely define each individual.

If \code{rec_method == 'thresh'}, \code{w_min_diam} must be defined. If \code{rec_method == 'extrap'}, \code{diam}, \code{min_size_class}, \code{min_diam_thresh} and \code{growth_percentile} must be defined.

Talbot et al. (2014) recommend a value of 0.86 for \code{growth_percentile} if \code{rec_method == 'extrap'}.
}
\examples{
data(bicuar)

obsRecGrowth(bicuar, "2019", "2021", w = "diam",
  group = "stem_id", census = "census_date",
  rec_method = "zero")

obsRecGrowth(bicuar, "2019", "2021", w = "diam",
  group = "stem_id", census = "census_date",
  rec_method = "thresh", w_min_diam = 5)

bicuar$agb_min <- runif(nrow(bicuar))
obsRecGrowth(bicuar, "2019", "2021", w = "agb",
  group = "stem_id", census = "census_date",
  rec_method = "thresh", w_min_diam = "agb_min")

obsRecGrowth(bicuar, "2019", "2021", w = "diam",
  group = "stem_id", census = "census_date",
  rec_method = "extrap", diam = "diam", min_size_class = c(5, 10),
  min_diam_thresh = 5, growth_percentile = 0.86)

}
