% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsSurGrowth.R
\name{obsSurGrowth}
\alias{obsSurGrowth}
\title{Calculate growth from individuals which survived between two censuses}
\usage{
obsSurGrowth(x, t0, tT, w, group, census, full = FALSE)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data}

\item{t0}{initial census of interval}

\item{tT}{final census of interval}

\item{w}{column name of value from which to calculate growth in \code{x}}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}

\item{full}{logical, if \code{TRUE} return a dataframe, otherwise a named vector}
}
\value{
Named vector or dataframe (if \code{full = TRUE}) of individual observed growth from individuals which survived.
}
\description{
This function acts on a datafame (\code{x}) of individuals (\code{group}) measured at two time points (\code{t0}, \code{tT}, \code{census}) to calculate the growth from individuals which survived, given a measure of individual size (\code{w}).
}
\details{
\code{group} can be a vector of multiple character strings which, when combined, uniquely define each individual.
}
\examples{
data(bicuar_clean)

obsSurGrowth(bicuar_clean, "2019", "2021", w = "diam", 
  group = c("plot_id", "stem_id"), census = "census_date")

}
