% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodKohyama.R
\name{prodKohyama}
\alias{prodKohyama}
\title{Estimate productivity according to Kohyama et al. (2019)}
\usage{
prodKohyama(x, t0, tT, w, group, census)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data from single plot}

\item{t0}{initial census of interval}

\item{tT}{final census of interval}

\item{w}{column name of value from which to calculate growth in \code{x}}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}
}
\value{
named list containing measures of productivity for the census interval:
\itemize{
\item \code{int} - census interval length
\item \code{t0} - year of initial census
\item \code{tT} - year of final census
\item \code{N0} - number of living stems in initial census
\item \code{NT} - number of living stems in final census
\item \code{Ns0} - number of surviving stems in initial census
\item \code{Nr0} - number of recruiting stems
\item \code{Nd0} - number of stems which died
\item \code{n0} - alternative method for number of stems in initial census, \code{Ns0 + Nd0}
\item \code{nT} - alternative method for number of stems in final census, \code{Ns0 + Nr0}
\item \code{dN} - observed net change in number of living stems \code{NT - N0}
\item \code{dN_ann} - observed net annual change in number of living stems \code{dN / int}
\item \code{B0} - initial biomass
\item \code{BT} - final biomass
\item \code{Bs0} - biomass of survivors in initial census
\item \code{Br0} - biomass of recruits in final census
\item \code{Bd0} - biomass lost to deaths
\item \code{dB} - observed net biomass change \code{BT - B0}
\item \code{dB_ann} - observed net annual biomass change \code{dB / int}
\item \code{W_max} - standardised maximum stem biomass for initial census (99th percentile of biomass)
\item \code{Bw} - period mean biomass \code{(BT-B0)/log(BT/B0)}
\item \code{Nw} - period mean abundance \code{(NT-N0)/log(NT/N0)}
\item \code{Bw_ann} - annual mean biomass \code{(BT-B0)/((BT/B0)^(1/int) - 1)/int}
\item \code{Nw_ann} - annual mean abundance \code{(NT-N0)/((NT/N0)^(1/int) - 1)/int}
\item \code{P} - instantaneous rate of production \code{(Bw * log(BT/Bs0))/int}
\item \code{L} - instantaneous rate of loss \code{(Bw * log(B0/Bs0))/int}
\item \code{P_simple} - simple rate of production \code{(B0-Bs0)/int}
\item \code{L_simple} - simple rate of loss \code{(BT-Bs0)/int}
\item \code{p} - intrinsic rate of production \verb{(log(BT/Bs0)/int}
\item \code{l} - intrinsic rate of loss \verb{(log(B0/Bs0)/int}
\item \code{r} - intrinsic rate of biomass change \verb{(log(BT/B0)/int}
\item \code{p_ann} - specific rate of production \code{(BT/B0)^(1/int) * (1 - (Bs0/BT)^(1/int))}
\item \code{l_ann} - specific rate of loss \code{1-(Bs0/B0)^(1/int)}
\item \code{lambda} - specific rate of biomass change \code{(BT/B0)^(1/int)}
\item \code{P_ann} - annual rate of production \code{Bw_ann * p_ann}
\item \code{L_ann} - annual rate of loss \code{Bw_ann * l_ann}
\item \code{Psimp} - simple rate of production, alternative method \code{(NT*BT - Ns0*B0) / int}
\item \code{Psimp_clark} - simple rate of production (Clark et al. 2001) \code{Ns0*(BT-B0) / int + Nr0*(BT-Bmin) / int}
}
}
\description{
This function acts on a datafame (\code{x}) of individuals (\code{group}) measured at two time points (\code{t0}, \code{tT}, \code{census}) to estimate rates of productivity and loss.
}
\details{
\code{group} can be a vector of multiple character strings which, when combined, uniquely define each individual.
}
\examples{
data(bicuar_clean)
bicuar_p1 <- bicuar_clean[bicuar_clean$plot_id == "ABG_5",]

prodKohyama(bicuar_p1, "2019", "2021", w = "diam",
  group = "stem_id", census = "census_date")

}
