% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodTalbot.R
\name{prodTalbot}
\alias{prodTalbot}
\title{Estimate productivity according to Talbot et al. (2014)}
\usage{
prodTalbot(
  x,
  t0,
  tT,
  w,
  diam,
  group,
  census,
  size_class = c(5, 10, 20, 30, 40, 50),
  min_size_class = c(5, 10),
  w_min_diam,
  rec_method = "zero",
  min_diam_thresh = NULL,
  growth_percentile = NULL
)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data from single plot}

\item{t0}{initial census of interval}

\item{tT}{final census of interval}

\item{w}{column name of value from which to calculate growth in \code{x}}

\item{diam}{column name of stem diameters in \code{x}}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}

\item{size_class}{numeric vector containing diameter size class cut points. The largest number will be open ended, e.g. 50+. Passed to \code{unobsMorGrowth()}}

\item{min_size_class}{vector of length two containing range of minimum diameter size class used to estimate median growth rate}

\item{w_min_diam}{column name or a vector of length 1 defining the estimated value of \code{w} in \code{x} at the minimum diameter threshold of the plot}

\item{rec_method}{character string describing method used to estimate growth of recruits. Can be \code{zero}, \code{thresh}, or \code{extrap}.}

\item{min_diam_thresh}{optional, minimum diameter threshold measured}

\item{growth_percentile}{optional, percentile of growth rate used to estimate growth rate of recruits for backward extrapolation}
}
\value{
named list containing measures of productivity:
\itemize{
\item \code{t0} - census date of initial census
\item \code{tT} - census date of final census
\item \code{int} - \code{tT - t0}
\item \code{sur_obs} - Observed growth of survivors
\item \code{rec_obs} - Observed growth contribution of recruits observed in final census
\item \code{mor_obs} - Observed loss from stems which died in census interval
\item \code{mor_unobs} - Estimated unobserved growth of stems which died in census interval
\item \code{rec_unobs} - Estimated unobserved growth of stems which recruited and died in census interval
\item \code{AGWP_obs} - Observed productivity: \code{sur_obs + rec_obs}
\item \code{AGWP_est} - Estimated productivity using CIC2, stem by stem approach: \code{AGWP_obs + mor_unobs + rec_unobs}
\item \code{AGWP_obs_ann} - \code{AGWP_obs / int}
\item \code{AGWP_est_ann} - \code{AGWP_est / int}
}
}
\description{
This function acts on a datafame (\code{x}) of individuals (\code{group}) measured at two time points (\code{t0}, \code{tT}, \code{census}) to estimate rates of producitivity.
}
\details{
\code{group} can be a vector of multiple character strings which, when combined, uniquely define each individual.

If \code{rec_method = 'thresh'}, \code{w_min_diam} must be defined. If \code{rec_method = 'extrap'}, \code{diam}, \code{min_size_class}, \code{min_diam_thresh} and \code{growth_percentile} must be defined.

Talbot et al. (2014) recommend a value of 0.86 for \code{growth_percentile} if \code{rec_method = 'extrap'}.
}
\examples{
data(bicuar_clean)
bicuar_p1 <- bicuar_clean[bicuar_clean$plot_id == "ABG_5",]

bicuar_p1$agb_min <- runif(nrow(bicuar_p1))
prodTalbot(bicuar_p1, "2019", "2021", w = "agb", diam = "diam",
  group = "stem_id", census = "census_date", w_min_diam = "agb_min", 
  rec_method = "zero")

prodTalbot(bicuar_p1, "2019", "2021", w = "agb", diam = "diam",
  group = "stem_id", census = "census_date", w_min_diam = "agb_min",
  size_class = seq(5, 20, 5), min_size_class = c(5, 10),
  rec_method = "zero")

prodTalbot(bicuar_p1, "2019", "2021", w = "diam", diam = "diam",
  group = "stem_id", census = "census_date", w_min_diam = 5, 
  rec_method = "thresh")

prodTalbot(bicuar_p1, "2019", "2021", w = "agb", diam = "diam",
  group = "stem_id", census = "census_date", w_min_diam = "agb_min",
  rec_method = "thresh")

prodTalbot(bicuar_p1, "2019", "2021", w = "agb", diam = "diam", 
  group = "stem_id", census = "census_date", w_min_diam = "agb_min", 
  rec_method = "extrap", min_diam_thresh = 5, growth_percentile = 0.86)
  
}
