% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unobsRec.R
\name{unobsRec}
\alias{unobsRec}
\title{Estimate unobserved number of recruits according to Talbot et al. (2014)}
\usage{
unobsRec(x, t0, tT, group, census)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data}

\item{t0}{initial census of interval}

\item{tT}{final census of interval}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}
}
\value{
\code{details_obs_sum(un = TRUE)} recruits.
}
\description{
This function acts on a datafame (\code{x}) of individuals (\code{group}) measured at two time points (\code{t0}, \code{tT}, \code{census}) to estimate the number of recruits which recruited and
died within a single census interval.
}
\details{
\code{group} can be a vector of multiple character strings which, when combined, uniquely define each individual.
}
\examples{
data(bicuar_clean)

unobsRec(bicuar_clean, "2019", "2021", 
  group = "stem_id", census = "census_date")

}
