% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unobsRecGrowth.R
\name{unobsRecGrowth}
\alias{unobsRecGrowth}
\title{Estimate growth of unobserved recruits between two censuses}
\usage{
unobsRecGrowth(x, t0, tT, w, group, census, diam, min_size_class, w_min_diam)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data}

\item{t0}{initial census of interval}

\item{tT}{final census of interval}

\item{w}{column name of value from which to calculate growth in \code{x}}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}

\item{diam}{column name of stem diameters in \code{x}}

\item{min_size_class}{vector of length two containing range of minimum diameter size class used to estimate median growth rate}

\item{w_min_diam}{column name or a vector of length 1 defining the estimated value of \code{w} in \code{x} at the minimum diameter threshold of the plot}
}
\value{
Numeric value of the sum of all unobserved growth from recruits which died.
}
\description{
This function acts on a datafame (\code{x}) of individuals (\code{group}) measured at two time points (\code{t0}, \code{tT}, \code{census}) to estimate the growth from individuals which recruited and died, given a measure of individual size (\code{w})
}
\details{
\code{group} can be a vector of multiple character strings which, when combined, uniquely define each individual.
}
\examples{
data(bicuar)

unobsRecGrowth(bicuar, "2019", "2021", w = "diam", group = "stem_id", 
  census = "census_date", diam = "diam", min_size_class = c(5, 10), 
  w_min_diam = 5)

}
