% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitalKohyama.R
\name{vitalKohyama}
\alias{vitalKohyama}
\title{Estimate vital rates according to Kohyama et al. (2018)}
\usage{
vitalKohyama(x, t0, tT, w = NULL, group, census, plot_area)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data from single plot}

\item{t0}{initial census of interval}

\item{tT}{final census of interval}

\item{w}{column name of value from which to calculate growth in \code{x} defaults to number of stems if \code{NULL}}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}

\item{plot_area}{plot area in hectares}
}
\value{
named list containing measures of vital rates for census interval:
\itemize{
\item \code{int} - census interval length, in years
\item \code{t0} - year of initial census = t0
\item \code{tT} - year of final census = tT
\item \code{N0} - Abundance of living stems in initial census
\item \code{NT} - Abundance of living stems in final census
\item \code{NST} - Abundance of surviving stems in final census
\item \code{nT} - Abundance of recruits in final census
\item \code{g} - Intrinsic rate of natural increase
\item \code{m} - Instantaneous per capita mortality rate
\item \code{r} - Instantaneous per capita recruitment rate
\item \code{l} - Finite rate of increase
\item \code{ma} - Per capita annual mortality rate
\item \code{ra} - Initial-density-based per capita annual recruitment rate
\item \code{raf} - Final-density-based per capita annual recruitment rate
\item \code{ras} - Survivor-density-based per capita annual recruitment rate
\item \code{raz} - Zero-mortality per capita annual recruitment rate
\item \code{M} - Instantaneous per area mortality rate
\item \code{R} - Instantaneous per area recruitment rate
\item \code{Ma} - Per area annual mortality rate
\item \code{Ra} - Per area annual recruitment rate
\item \code{Ras} - Per area annual recruitment rate with first year deaths
}
}
\description{
This function acts on a datafame (\code{x}) of individuals (\code{group}) measured at two time points (\code{t0}, \code{tT}, \code{census}) to estimate rates of recruitment, mortality, and net
population change.
}
\details{
\code{group} can be a vector of multiple character strings which, when combined, uniquely define each individual.
}
\examples{
data(bicuar)

vitalKohyama(bicuar, "2019", "2021", group = "stem_id", 
  census = "census_date", plot_area = 1)

vitalKohyama(bicuar, "2019", "2021", group = "stem_id", w = "agb",
  census = "census_date", plot_area = 1)

}
