% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodKohyama.R
\name{prodKohyama}
\alias{prodKohyama}
\title{Estimate productivity according to Kohyama et al. (2019)}
\usage{
prodKohyama(x, t0, tT, w, group, census)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data from single plot}

\item{t0}{initial census of interval}

\item{tT}{final census of interval}

\item{w}{column name of value from which to calculate growth in \code{x}}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}
}
\value{
named list containing measures of productivity for the census interval:
\itemize{
\item \code{int} - census interval length
\item \code{t0} - year of initial census
\item \code{tT} - year of final census
\item \code{N0} - number of living stems in initial census
\item \code{NT} - number of living stems in final census
\item \code{Ns0} - number of surviving stems in initial census
\item \code{Nr0} - number of recruiting stems
\item \code{Nd0} - number of stems which died
\item \code{B0} - initial biomass
\item \code{BT} - final biomass
\item \code{dB} - biomass change
\item \code{dB_ann} - annual biomass change
\item \code{Bs0} - biomass of survivors in initial census
\item \code{Br0} - biomass of recruits in final census
\item \code{Bd0} - biomass lost to deaths
\item \code{W_max} - standardised maximum stem biomass for initial census (99th percentile of biomass)
\item \code{r_turn} - instantaneous recruitment rate
\item \code{m_turn} - instantaneous mortality rate
\item \code{p_turn} - instantaneous production rate
\item \code{l_turn} - instantaneous loss rate
\item \code{Nw} - period mean abundance (Kohyama et al. 2019 - Eq10)
\item \code{Nw_ann} - annual mean abundance (Kohyama et al. 2019 - Eq14)
\item \code{Bw} - period mean biomass (Kohyama et al. 2019 - Eq10)
\item \code{Bw_ann} - annual mean biomass (Kohyama et al. 2019 - Eq14)
\item \code{P_simple} - simple rate of production \code{(B0-Bs0)/int}
\item \code{L_simple} - simple rate of loss \code{(BT-Bs0)/int}
\item \code{Bw_simple} - simple mean biomass \code{(B0+BT)/2}
\item \code{P_ann} - annual rate of production (Kohyama et al. 2019 Eq12)
\item \code{L_ann} - annual rate of loss (Kohyama et al. 2019 Eq13)
\item \code{P} - instantaneous rate of production
\item \code{L} - instantaneous rate of loss
\item \code{Bwk} - alternative measure of period mean biomass, to check consistency in calculation with \code{Bw}
\item \code{Psimp} - simple rate of production, alternative method
\item \code{Psimp_clark} - simple rate of production (Clark et al. 2001)
}
}
\description{
This function acts on a datafame (\code{x}) of individuals (\code{group}) measured at two time points (\code{t0}, \code{tT}, \code{census}) to estimate rates of productivity and loss.
}
\details{
\code{group} can be a vector of multiple character strings which, when combined, uniquely define each individual.
}
\examples{
data(bicuar_clean)
bicuar_p1 <- bicuar_clean[bicuar_clean$plot_id == "ABG_5",]

prodKohyama(bicuar_p1, "2019", "2021", w = "diam",
  group = "stem_id", census = "census_date")

}
