% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unobsMorGrowth.R
\name{unobsMorGrowth}
\alias{unobsMorGrowth}
\title{Estimate growth of individuals which died between two censues}
\usage{
unobsMorGrowth(
  x,
  t0,
  tT,
  w,
  group,
  census,
  diam,
  size_class = c(5, 10, 20, 30, 40, 50)
)
}
\arguments{
\item{x}{dataframe of SEOSAW stem data}

\item{t0}{initial census of interval}

\item{tT}{final census of interval}

\item{w}{column name of value from which to calculate growth in \code{x}}

\item{group}{column name(s) defining individual IDs in \code{x}}

\item{census}{column name of census dates in \code{x}}

\item{diam}{column name of stem diameters in \code{x}}

\item{size_class}{numeric vector containing diameter size class cut points. The largest number will be open ended, e.g. 50+.}
}
\value{
Numeric value of the sum of all unobserved growth from individuals which died.
}
\description{
This function acts on a datafame (\code{x}) of individuals (\code{group}) measured at two time points (\code{t0}, \code{tT}, \code{census}) to estimate the growth from individuals which died, given a measure of individual size (\code{w})
}
\details{
\code{group} can be a vector of multiple character strings which, when combined, uniquely define each individual.
}
\examples{
data(bicuar)

unobsMorGrowth(bicuar, "2019", "2021", w = "diam", 
  group = "stem_id", census = "census_date", diam = "diam")

}
